Perl-Support: Improve perldoc integration
- Change the way the perldoc integration reports errors. The error messages
  printed by perldoc are now listed in the buffer. See below.
- Minor changes.

Under Ubuntu, if the perldoc package is not installed, Perl still comes with
a "dummy" version of perldoc, which fails and reports:
  "You need to install the perl-doc package to use this program."
Thus, with the old implementation, it seemed like no help has been found for
the keyword. To solve this problem, we now list the output of the failed
perldoc calls in the buffer.